+++ WWW/Library/Implementation/HTString.c
@@ -1,5 +1,5 @@
 /*
- * $LynxId: HTString.c,v 1.55 2008/12/31 22:03:25 tom Exp $
+ * $LynxId: HTString.c,v 1.56 2009/01/01 16:57:48 tom Exp $
  *
  *	Case-independent string comparison		HTString.c
  *
@@ -604,9 +604,9 @@ typedef enum {
     Format
 } PRINTF;
 
-#define VA_INTGR(type) ival = va_arg((*ap), type)
-#define VA_FLOAT(type) fval = va_arg((*ap), type)
-#define VA_POINT(type) pval = (char *)va_arg((*ap), type)
+#define VA_INTGR(type) ival = (int)    va_arg((*ap), type)
+#define VA_FLOAT(type) fval = (double) va_arg((*ap), type)
+#define VA_POINT(type) pval = (char *) va_arg((*ap), type)
 
 #define NUM_WIDTH 10		/* allow for width substituted for "*" in "%*s" */
 		/* also number of chars assumed to be needed in addition
